+++ src/barchart_ui.c
@@ -29,7 +29,7 @@ static gboolean barchart_scale (gboolean button1_p, gboolean button2_p,
 
 void barchart_set_initials (splotd * sp, GGobiStage * d);
 void barchart_allocate_structure (splotd * sp, GGobiStage * d);
-extern void barchart_set_breakpoints (gfloat width, splotd * sp,
+extern void barchart_set_breakpoints (gdouble width, splotd * sp,
                                       GGobiStage * d);
 
 static void
@@ -292,18 +292,18 @@ barchart_scale (gboolean button1_p, gboolean button2_p, splotd * sp)
     fcoords pts1, pts2;
 
     if (bsp->bar->anchor_drag) {
-      gfloat scale_y;
+      gdouble scale_y;
       icoords scr;
 
       if (idy != 0) {
         gboolean set_anchor = TRUE;
-        gfloat offset_old = bsp->bar->offset;
+        gdouble offset_old = bsp->bar->offset;
         gint pmid_old = sp->pmid.y;
 
         scr.x = scr.y = 0;
         scale_y = sp->scale.y;
         scale_y /= 2;
-        sp->iscale.y = (gfloat) sp->max.y * scale_y;
+        sp->iscale.y = (gdouble) sp->max.y * scale_y;
 
         klass->screen_to_tform (cpanel, sp, &scr, &pts1, gg);
         sp->pmid.y -= (dy * (gdouble) PRECISION1 / sp->iscale.y);
@@ -336,7 +336,7 @@ barchart_scale (gboolean button1_p, gboolean button2_p, splotd * sp)
     else {                      /* if (bsp->bar->width_drag) */
 
       if (idy != 0) {
-        gfloat width, oldwidth;
+        gdouble width, oldwidth;
 
         klass->screen_to_tform (cpanel, sp, &sp->mousepos_o, &pts1, gg);
         klass->screen_to_tform (cpanel, sp, &sp->mousepos, &pts2, gg);
