+++ src/svd.h
@@ -1,6 +1,6 @@
 #ifndef GGOBI_SVD_H
 #define GGOBI_SVD_H
 
-gint       dsvd (gdouble **a, gint m, gint n, gfloat *w, gdouble **v);
+gint       dsvd (gdouble **a, gint m, gint n, gdouble *w, gdouble **v);
 
 #endif
